feat: update Order and PriceCalculator to use BigDecimal for totalPrice calculations

+++ monolith/src/main/java/com/optivem/atddaccelerator/template/monolith/common/Order.java
@@ -1,15 +1,16 @@
 package com.optivem.atddaccelerator.template.monolith.common;
 
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 public class Order {
     private String orderNumber;
     private long productId;
     private int quantity;
-    private double totalPrice;
+    private BigDecimal totalPrice;
 
-    public Order(String orderNumber, long productId, int quantity, double totalPrice) {
+    public Order(String orderNumber, long productId, int quantity, BigDecimal totalPrice) {
         this.orderNumber = orderNumber;
         this.productId = productId;
         this.quantity = quantity;
